perf(kotlin): Stop Vite from reloading all Kotlin files
We previously used the Gradle built-in `Sync` task. The `Sync` task always rewrites the files, even if they haven't changed. Vite detects this change and thinks the entire codebase has changed.

It turns out we're not the first ones to have this problem because the Kotlin JS plugin has its task, `DefaultIncrementalSyncTask`, to deal with this.

In this commit, we reuse the Kotlin JS' task to incrementally copy the files. As a result, Vite only sees one or two files changing, and reloads much faster.